+++ src/app/components/file-browser/fb-table-row.js
@@ -9,7 +9,7 @@
 
 import Component from '@ember/component';
 import { reads, not } from '@ember/object/computed';
-import { equal, raw, or, conditional, isEmpty } from 'ember-awesome-macros';
+import { equal, raw, or, and, array, conditional, isEmpty } from 'ember-awesome-macros';
 import { get, computed, getProperties, observer } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { later, cancel, scheduleOnce } from '@ember/runloop';
@@ -142,6 +142,18 @@ export default Component.extend(I18n, FastDoubleClick, {
   qosViewForbidden: false,
 
   /**
+   * @virtual
+   * @type {Boolean}
+   */
+  datasetsViewForbidden: false,
+
+  /**
+   * Name of icon to indicate that some property in tag is inhertied from ancestor
+   * @type {String}
+   */
+  inheritedIcon: 'arrow-long-up',
+
+  /**
    * Time in ms when the touch should be treated as a hold
    * @type {number}
    */
@@ -176,11 +188,27 @@ export default Component.extend(I18n, FastDoubleClick, {
 
   fileNameSuffix: reads('fileNameParser.suffix'),
 
+  /**
+   * Text for QoS tag tooltip, when cannot open QoS modal
+   * @type {ComputedProperty<SafeString>}
+   */
   hintQosViewForbidden: computed(function hintQosForbidden() {
     return insufficientPrivilegesMessage({
       i18n: this.get('i18n'),
       modelName: 'space',
       privilegeFlag: 'space_view_qos',
+    });
+  }),
+
+  /**
+   * Text for dataset tag tooltip, when cannot open datasets modal
+   * @type {ComputedProperty<SafeString>}
+   */
+  hintDatasetsViewForbidden: computed(function hintDatasetsViewForbidden() {
+    return insufficientPrivilegesMessage({
+      i18n: this.get('i18n'),
+      modelName: 'space',
+      privilegeFlag: 'space_view',
     });
   }),
 
@@ -386,13 +414,80 @@ export default Component.extend(I18n, FastDoubleClick, {
 
   hasMetadata: reads('file.hasMetadata'),
 
-  hasEffQos: reads('file.hasEffQos'),
+  effQosMembership: reads('file.effQosMembership'),
 
-  hasDirectQos: reads('file.hasDirectQos'),
+  showQosTag: and(
+    not('previewMode'),
+    not('isSymlink'),
+    array.includes(raw(['ancestor', 'direct']), 'effQosMembership')
+  ),
 
   hasAcl: equal('file.activePermissionsType', raw('acl')),
 
+  effDatasetMembership: reads('file.effDatasetMembership'),
+
   hardlinksCount: or('file.hardlinksCount', raw(1)),
+
+  /**
+   * If true, should display dataset tag
+   * @type {ComputedProperty<Boolean>}
+   */
+  showDatasetTag: and(
+    not('previewMode'),
+    not('isSymlink'),
+    array.includes(raw(['ancestor', 'direct']), 'effDatasetMembership')
+  ),
+
+  /**
+   * @type {ComputedProperty<Boolean>}
+   */
+  dataIsProtected: and(
+    'showDatasetTag',
+    'file.dataIsProtected'
+  ),
+
+  /**
+   * @type {ComputedProperty<Boolean>}
+   */
+  metadataIsProtected: and(
+    'showDatasetTag',
+    'file.metadataIsProtected'
+  ),
+
+  /**
+   * @type {ComputedProperty<Boolean>}
+   */
+  hasAnyProtectionFlag: or('metadataIsProtected', 'dataIsProtected'),
+
+  /**
+   * Content for protection tag tooltip
+   * @type {ComputedProperty<SafeString>}
+   */
+  protectionFlagsInfo: computed(
+    'typeText',
+    'metadataIsProtected',
+    'dataIsProtected',
+    function protectionFlagsInfo() {
+      const {
+        typeText,
+        metadataIsProtected,
+        dataIsProtected,
+      } = this.getProperties('typeText', 'metadataIsProtected', 'dataIsProtected');
+      let translationKey;
+      if (dataIsProtected && metadataIsProtected) {
+        translationKey = 'both';
+      } else if (dataIsProtected) {
+        translationKey = 'data';
+      } else if (metadataIsProtected) {
+        translationKey = 'metadata';
+      }
+      if (translationKey) {
+        return this.t(`protectionFlagsInfo.${translationKey}`, { fileType: typeText });
+      } else {
+        return '';
+      }
+    }
+  ),
 
   loadingOnIconTransitionObserver: observer(
     'isLoadingOnIcon',